requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
# requirements.txt
torch @ https://download.pytorch.org/whl/cu128/torch
torchvision @ https://download.pytorch.org/whl/cu128/torchvision
torchaudio @ https://download.pytorch.org/whl/cu128/torchaudio
pyqt5 
qt-material
opencv-python
requests
grpcio
protobuf
pyaudio